역학조사서 삭제 endpoint 작성(진행중), DynamicQuery.kt 수정

+++ src/main/kotlin/org/sbas/dtos/InfoPtDto.kt
@@ -45,7 +45,7 @@ fun InfoPtDto.toEntity(): InfoPt {
 @NoArg
 data class NewsScoreParam(
     var breath: Int,  // 분당호흡수
-    var spo2: Int,    // 산소포화도
+    var spo2: Double,    // 산소포화도
     var o2Apply: String, // 산소투여여부
     var sbp: Int,     // 수축기혈압
     var pulse: Int,   // 맥박
